Run check_env_vars task first in pipeline
Currently the `check_env_vars` task is isolated with no dependencies to
the rest of the pipeline. Yes, if this task does fail the DAG run will
ultimately fail, but so will the `get_query_ids` task as well (for
effectively the same reason). The `check_env_vars` task should act as a
gate for the pipeline, and if the task is still useful to function this
way, it should be the first task in the pipeline.