Add end-time pruning when accessing Snowflake views
There are times when the query performance against the Snowflake
ACCOUNT_USAGE views tanks -- for the same number of queries to retrieve
information, the query can take seconds or 2-4+ minutes. Add some sort
of time-range filter to the views ensures that the queries perform
consistently and timely.

The proposed changed is the set a `end_time` threshold of 24 hours prior
to the end interval that's used to retrieve the query IDs initially
(which effectively becomes 30 hours ago based on the current Snowflake
data latency).